backend/requirements.txt
Django==4.2
mysqlclient==2.0
gunicorn==20.1
python-dotenv
PyMySQL==1.1.0
mysql-connector-python
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
drf-spectacular==0.27.2
django-cors-headers==4.5.0
Pillow
feedparser
yfinance
requests
beautifulsoup4
Faker==18.11.1
pandas
seaborn
matplotlib
sklearn
imblearn
xgboost
tensorflow
category_encoders
transformers
mistralai
google-generativeai